add user validation to delete/update bid (context user must match)

+++ server/schemas/resolvers.js
@@ -91,6 +91,7 @@ const resolvers = {
           ).populate('auctions');
         return updatedUser ;
       }
+      throw new AuthenticationError('Incorrect credentials');
     },
     addBid: async (parent, { input }, context) => {
       if( context.user ){
@@ -116,52 +117,65 @@ const resolvers = {
           ).populate('bids');
 
         return updatedAuction
-
       }
+      throw new AuthenticationError('Incorrect credentials');
     },
     deleteBid: async (parent, { _id }, context) => {
       if( context.user ){
           // get data for current bid
         const currentBid = await Bid.findOne( { _id } )
 
-          // delete bid
-        const bid = await Bid.deleteOne( { _id } );  
+          // if context user matches the owner of the bid, allow delete
+        if( currentBid.userId == context.user._id) {
 
-          // pull bid from user array
-        const updatedUser  = await User.findOneAndUpdate(
-          {_id: currentBid.userId },
-          { $pull: { bids: _id } },
-          { new: true, runValidators: true }
-          ).populate('bids');
+            // delete bid
+          const bid = await Bid.deleteOne( { _id } );  
 
-            // pull bid from autcion array
-        const updatedAuction  = await Auction.findOneAndUpdate(
-          {_id: currentBid.auctionId},
-          { $pull: { bids: _id } },
-          { new: true, runValidators: true }
-          ).populate('bids');
+            // pull bid from user array
+          const updatedUser  = await User.findOneAndUpdate(
+            {_id: currentBid.userId },
+            { $pull: { bids: _id } },
+            { new: true, runValidators: true }
+            ).populate('bids');
 
-        return updatedAuction
+              // pull bid from autcion array
+          const updatedAuction  = await Auction.findOneAndUpdate(
+            {_id: currentBid.auctionId},
+            { $pull: { bids: _id } },
+            { new: true, runValidators: true }
+            ).populate('bids');
+
+          return updatedAuction
+        }
+        return currentBid
       }
+      throw new AuthenticationError('Incorrect credentials');
     },
     updateBid: async (parent, { _id, maxBid, increment} , context) => {
       if( context.user ){
           // get data for current bid
         const currentBid = await Bid.findOne( { _id } )
 
-          // verify that maxBid and increment are not being reduced from current settings, otherwise continue with current
-        maxBid = maxBid <= currentBid.maxBid ? currentBid.maxBid : maxBid
-        increment = increment <= currentBid.increment ? currentBid.increment : increment
+          // if context user matches the owner of the bid, allow update
+        if( currentBid.userId == context.user._id) {
+            // verify that maxBid and increment are not being reduced from current settings, otherwise continue with current
+          maxBid = maxBid <= currentBid.maxBid ? currentBid.maxBid : maxBid
+          increment = increment <= currentBid.increment ? currentBid.increment : increment
+    
+            // update bid 
+          const updatedBid  = await Bid.findOneAndUpdate(
+            { _id },
+            { maxBid, increment },
+            { new: true, runValidators: true }
+            )
 
-          // update bid 
-        const updatedBid  = await Bid.findOneAndUpdate(
-          { _id },
-          { maxBid, increment },
-          { new: true, runValidators: true }
-          )
+          return updatedBid
+        }
+          // else return the original bid
+        return currentBid
 
-        return updatedBid
       }
+      throw new AuthenticationError('Incorrect credentials');
     },
   }
 }
